Replaced double -> float in harmonic.cpp. Removed std::unique_ptr from LCD, using now normal class member.

+++ harmonic.cpp
@@ -2,12 +2,12 @@
 
 // Solve A..E with least squares method from 8 given datapoints
 HarmonicCoeffs computeHarmonicCoeffs(const float* dev_deg) {
-  double MTM[5][5] = {0};
-  double MTy[5] = {0};
+  float MTM[5][5] = {0};
+  float MTy[5] = {0};
   for (int i=0; i<8; i++) {
-    double th = headings_deg[i] * M_PI/180.0;
-    double v[5] = {1.0, sin(th), cos(th), sin(2*th), cos(2*th)};
-    double y = dev_deg[i];
+    float th = headings_deg[i] * M_PI/180.0;
+    float v[5] = {1.0, sin(th), cos(th), sin(2*th), cos(2*th)};
+    float y = dev_deg[i];
     for (int r=0; r<5; r++) {
       MTy[r] += v[r]*y;
       for (int c=0; c<5; c++) MTM[r][c] += v[r]*v[c];
@@ -15,7 +15,7 @@ HarmonicCoeffs computeHarmonicCoeffs(const float* dev_deg) {
   }
 
   // Gauss elimination
-  double A[5][6];
+  float A[5][6];
   for (int r=0; r<5; r++) {
     for (int c=0; c<5; c++) A[r][c] = MTM[r][c];
     A[r][5] = MTy[r];
@@ -25,13 +25,13 @@ HarmonicCoeffs computeHarmonicCoeffs(const float* dev_deg) {
     for (int r=i+1; r<5; r++)
       if (fabs(A[r][i]) > fabs(A[piv][i])) piv = r;
     if (piv != i)
-      for (int c=i; c<6; c++) { double tmp = A[i][c]; A[i][c] = A[piv][c]; A[piv][c] = tmp; }
-    double diag = A[i][i];
+      for (int c=i; c<6; c++) { float tmp = A[i][c]; A[i][c] = A[piv][c]; A[piv][c] = tmp; }
+    float diag = A[i][i];
     if (fabs(diag) < 1e-9) continue;
     for (int c=i; c<6; c++) A[i][c] /= diag;
     for (int r=0; r<5; r++) {
       if (r == i) continue;
-      double f = A[r][i];
+      float f = A[r][i];
       for (int c=i; c<6; c++) A[r][c] -= f * A[i][c];
     }
   }
